Dark mode toggle now sits to the right of the terms & conditions at bottom. Could we make this more descriptive/clear?
