+++ src/components/app/header/index.tsx
@@ -8,7 +8,6 @@ import {isEmpty} from 'lodash'
 import Feedback from 'components/feedback-input'
 import useBreakpoint from 'utils/breakpoints'
 import {useRouter} from 'next/router'
-import {useTheme} from 'next-themes'
 import useCio from 'hooks/use-cio'
 import {Form, Formik} from 'formik'
 import ProjectClubCTA from 'components/survey/project-club'
@@ -139,7 +138,6 @@ const Header: FunctionComponent = () => {
                 </span>
               </a>
             </Link>
-            <DarkModeToggle />
           </div>
         ) : (
           <div className={className}>
@@ -157,7 +155,6 @@ const Header: FunctionComponent = () => {
                 Sign in
               </a>
             </Link>
-            <DarkModeToggle />
           </div>
         )}
       </div>
@@ -253,59 +250,6 @@ const SearchBar = () => {
         )
       }}
     </Formik>
-  )
-}
-
-const DarkModeToggle = () => {
-  const [mounted, setMounted] = React.useState(false)
-  const {subscriber, cioIdentify} = useCio()
-  const {theme, setTheme} = useTheme()
-  React.useEffect(() => setMounted(true), [])
-
-  return (
-    <button
-      aria-label="Toggle Dark Mode"
-      type="button"
-      className="rounded p-3 h-10 w-10"
-      onClick={() => {
-        const nextTheme = theme === 'dark' ? 'light' : 'dark'
-        setTheme(nextTheme)
-        track(`toggled dark mode`, {
-          mode: nextTheme,
-        })
-        if (subscriber) {
-          cioIdentify(subscriber.id, {
-            theme_preference: nextTheme,
-          })
-        }
-      }}
-    >
-      {mounted && (
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          viewBox="0 0 24 24"
-          fill="currentColor"
-          stroke="currentColor"
-          className="h-4 w-4 text-gray-800 dark:text-gray-200"
-        >
-          {theme === 'dark' ? (
-            <path
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              strokeWidth={2}
-              d="M12 3v1m0 16v1m9-9h-1M4 12H3m15.364 6.364l-.707-.707M6.343 6.343l-.707-.707m12.728 0l-.707.707M6.343 17.657l-.707.707M16 12a4 4 0 11-8 0 4 4 0 018 0z"
-            />
-          ) : (
-            <path
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              strokeWidth={2}
-              d="M20.354 15.354A9 9 0 018.646 3.646 9.003 9.003 0 0012 21a9.003 9.003 0 008.354-5.646z"
-            />
-          )}
-        </svg>
-      )}
-    </button>
   )
 }
